BUG: Fix labelled and misplaced image references
Fix labelled and misplaced image refereces.

Labelled image references of the type
`![Xception architecture](/deep-learning/images/xception/architecture.png)`
are fixed thanks to the improved `sed` command.

Misplaced image references are those that were referencing a different
category to the one the review article was located in: e.g. the
`2018-09-14-morphnet.md` post was located in the `pruning-acceleration`
folder/category, but its images were located in the `deep-learning`
folder/category. Thus, the `for` loop was not matching the latter while
modifying the references of the `pruning-acceleration` folder posts.

A second execution of the script fixes this while modifying the
references of the `deep-learning` folder posts.

Lines containing multiple references are also fixed, either thanks to the
improvement in the `sed` command or thanks to executing the script more
than once.