BUG: Fix the hook
Fix the hook: the hook was actually not being called, so the previous
procedure is changed to explicitly call the `git config` command to tell
git to find the hooks in the specified folder.

The `.hooks-config.bash` script must be called explicitly to set up the
hooks.

Add the executable flag to the `pre-commit` hook file.

Other comments:
- Add the modified files to the files to be checked.
- Call the scripts as bash executable scripts.
- Avoid calling the scripts if no Markdown were added or modified.