BUG: Exclude the top-level Markdown files from the hook
Exclude the top-level Markdown files from the hook.

The pre-commit hook inspects any added or modified Markdown files,
including those in the top-level (e.g. the `README.md`, `how-to.md`,
etc.), so any tag or cross-reference in those files (e.g. those in
the `how-to.md` example) will also be considered for inspection.
Evidently, the ones in the `how-to.md` do not need to exist, but the
hook will identify them as missing if the file has been modified and
added to the commit.

This commit excludes the top-level files from the files to be considered
by the pre-commit hook.