ENH: Refactor the regex to find cross-references to deal with more cases
Refactor the regex to find cross-references to deal with more cases.

Characters after the closing brackets were being considered as part of
the cross-referenced file (e.g. commas or display properties such as
`{:align="middle"}`, `{: width="49%"}`, `{: style="width: 365px"}`,
CRLFs (fixed in #376), or because there were multiple brackets, such as
the alt text in
https://github.com/vitalab/vitalab.github.io/blob/master/article/_posts/2019-02-21-PlaNet.md,
and the regex was considering only the first and last brackets).

So the proposed regex
- Splits the target cross-references in more groups so as to consider
  the characters after the target filename as belonging to another
group.
- For some reason the command with the `-e` flag did not like the regex,
  and I had to switch to the `-r` flag, which apparently makes necessary
to locate the opening and closing brackets around the target filename so
that the variable that will host the filename contains no brackets.